requirements.txt
numpy >= 1.23.5
matplotlib >= 3.7.1
pennylane >= 0.35.1
torch >= 2.0.1
tqdm >= 4.65.0
pyyaml >= 6.0
